backend/requirements.txt
fastapi
uvicorn
python-dotenv
snowflake-connector-python
pandas
matplotlib
langchain
langgraph
tiktoken
litellm
google-generativeai
openai
anthropic
pinecone
sentence-transformers
yahooquery
tavily-python
langchain-community
google-search-results==2.4.2

